refactor: integrate auth_api
