+++ src/rpc/auth_api.rs
@@ -2,13 +2,28 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 use crate::auth::*;
+use crate::lotus_json::lotus_json_with_self;
 use crate::lotus_json::LotusJson;
 use crate::rpc::error::JsonRpcError;
 use crate::rpc::Ctx;
-use crate::rpc_api::auth_api::*;
 use anyhow::Result;
+use chrono::Duration;
 use fvm_ipld_blockstore::Blockstore;
 use jsonrpsee::types::Params;
+use serde::{Deserialize, Serialize};
+use serde_with::{serde_as, DurationSeconds};
+
+pub const AUTH_NEW: &str = "Filecoin.AuthNew";
+#[serde_as]
+#[derive(Deserialize, Serialize)]
+pub struct AuthNewParams {
+    pub perms: Vec<String>,
+    #[serde_as(as = "DurationSeconds<i64>")]
+    pub token_exp: Duration,
+}
+lotus_json_with_self!(AuthNewParams);
+
+pub const AUTH_VERIFY: &str = "Filecoin.AuthVerify";
 
 /// RPC call to create a new JWT Token
 pub async fn auth_new<DB: Blockstore>(
