+++ src/rpc_client/auth_ops.rs
@@ -1,7 +1,7 @@
 // Copyright 2019-2024 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use crate::rpc_api::auth_api::*;
+use crate::rpc::auth_api::*;
 use chrono::Duration;
 
 use super::{ApiInfo, JsonRpcError, RpcRequest};
